+++ src/aui-datatype/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-2061](https://issues.liferay.com/browse/AUI-2061) Low performance on countDays method when using huge dates.
 
 ## [3.0.3](https://github.com/liferay/alloy-ui/releases/tag/3.0.3)